Rearrange stack and utility
