+++ src/utility.h
@@ -73,43 +73,16 @@ typedef struct Bbinding_prototype {
   Data *data;
 } Binding;
 
-Binding *new_binding(char *sig, Data *data) {
-  Binding *res = NEW(Binding);
-  res->sig = sig;
-  res->data = data;
-  return res;
-}
+Binding *new_binding(char *sig, Data *data);
 
-Data *new_primitive_data(Primitive val) {
-  Data *res = NEW(Data);
-  res->type = D_PRIMITIVE;
-  res->data->primitive = NEW(Primitive);
-  *res->data->primitive = val;
-  return res;
-}
+Data *new_primitive_data(Primitive val);
 
-Data *new_closure_data(struct list *params, struct cmd *body) {
-  Data *res = NEW(Data);
-  res->type = D_CLOSURE;
-  res->data->closure = NEW(Closure);
-  res->data->closure->params = params;
-  res->data->closure->body = body;
-  return res;
-}
+Data *new_closure_data(struct list *params, struct cmd *body);
 
-Binding *new_empty_binding(char *sig) {
-  Data *data = NEW(Data);
-  data->type = D_EMPTY;
-  data->data = NULL;
-  return new_binding(sig, data);
-}
+Binding *new_empty_binding(char *sig);
 
-Binding *new_primitive_binding(char *sig, Primitive val) {
-  return new_binding(sig, new_primitive_data(val));
-}
+Binding *new_primitive_binding(char *sig, Primitive val);
 
-Binding *new_closure_binding(char *sig, struct list *params, struct cmd *body) {
-  return new_binding(sig, new_closure_data(params, body));
-}
+Binding *new_closure_binding(char *sig, struct list *params, struct cmd *body);
 
 #endif
